+++ src/main/java/inu/codin/codin/domain/post/domain/comment/repository/CommentRepository.java
@@ -1,0 +1,13 @@
+package inu.codin.codin.domain.post.domain.comment.repository;
+
+import inu.codin.codin.domain.post.domain.comment.entity.CommentEntity;
+import jakarta.validation.constraints.NotBlank;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.List;
+
+public interface CommentRepository extends MongoRepository<CommentEntity, String> {
+    List<CommentEntity> findByPostId(String postId);
+
+    int countByPostId(@NotBlank String postId);
+}
