+++ src/main/java/inu/codin/codin/domain/post/domain/comment/repository/ReplyRepository.java
@@ -1,0 +1,11 @@
+package inu.codin.codin.domain.post.domain.comment.repository;
+
+
+import inu.codin.codin.domain.post.domain.comment.entity.ReplyEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.util.List;
+
+public interface ReplyRepository extends MongoRepository<ReplyEntity, String> {
+    List<ReplyEntity> findByCommentId(String commentId);
+}
