+++ src/main/java/inu/codin/codin/domain/post/domain/like/entity/LikeType.java
@@ -1,0 +1,16 @@
+package inu.codin.codin.domain.post.domain.like.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum LikeType {
+    post("게시물"),
+    comment("댓글"),
+    reply("대댓글");
+
+    private final String description;
+
+    LikeType(String description) {
+        this.description = description;
+    }
+}
