+++ src/main/java/inu/codin/codin/domain/post/domain/like/exception/InvalidLikeTypeException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.post.domain.like.exception;
+
+public class InvalidLikeTypeException extends RuntimeException {
+    public InvalidLikeTypeException(String message) {
+        super(message);
+    }
+}
