+++ src/main/java/inu/codin/codin/domain/post/domain/like/exception/LikeCreateFailException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.post.domain.like.exception;
+
+public class LikeCreateFailException extends RuntimeException{
+    public LikeCreateFailException(String message) {
+        super(message);
+    }
+}
