+++ src/main/java/inu/codin/codin/domain/post/domain/like/exception/LikeRemoveFailException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.post.domain.like.exception;
+
+public class LikeRemoveFailException extends RuntimeException {
+    public LikeRemoveFailException(String message) {
+        super(message);
+    }
+}
