+++ src/main/java/inu/codin/codin/domain/post/domain/scrap/exception/ScrapCreateFailException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.post.domain.scrap.exception;
+
+public class ScrapCreateFailException extends RuntimeException {
+    public ScrapCreateFailException(String message) {
+        super(message);
+    }
+}
