+++ src/main/java/inu/codin/codin/domain/post/domain/scrap/exception/ScrapRemoveFailException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.post.domain.scrap.exception;
+
+public class ScrapRemoveFailException extends RuntimeException {
+    public ScrapRemoveFailException(String message) {
+        super(message);
+    }
+}
