--- src/main/java/inu/codin/codin/domain/post/like/exception/InvalidLikeTypeException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.post.like.exception;
-
-public class InvalidLikeTypeException extends RuntimeException {
-    public InvalidLikeTypeException(String message) {
-        super(message);
-    }
-}
