--- src/main/java/inu/codin/codin/domain/post/like/exception/LikeCreateFailException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.post.like.exception;
-
-public class LikeCreateFailException extends RuntimeException{
-    public LikeCreateFailException(String message) {
-        super(message);
-    }
-}
