--- src/main/java/inu/codin/codin/domain/post/like/exception/LikeRemoveFailException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.post.like.exception;
-
-public class LikeRemoveFailException extends RuntimeException {
-    public LikeRemoveFailException(String message) {
-        super(message);
-    }
-}
