--- src/main/java/inu/codin/codin/domain/post/scrap/exception/ScrapCreateFailException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.post.scrap.exception;
-
-public class ScrapCreateFailException extends RuntimeException {
-    public ScrapCreateFailException(String message) {
-        super(message);
-    }
-}
