--- src/main/java/inu/codin/codin/domain/post/scrap/exception/ScrapRemoveFailException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.post.scrap.exception;
-
-public class ScrapRemoveFailException extends RuntimeException {
-    public ScrapRemoveFailException(String message) {
-        super(message);
-    }
-}
