+++ src/main/java/inu/codin/codin/infra/redis/RedisRecoverSyncScheduler.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.infra.redis;
 
-import inu.codin.codin.domain.post.like.LikeService;
-import inu.codin.codin.domain.post.scrap.ScrapService;
+import inu.codin.codin.domain.post.domain.like.LikeService;
+import inu.codin.codin.domain.post.domain.scrap.ScrapService;
 import inu.codin.codin.infra.redis.exception.RedisUnavailableException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
